demo/requirements.txt
django-tinymce
django-filebrowser-no-grappelli
